fix(ci): line-limit check — executor helpers exceeded the module cap
