+++ src/awf/control/executor/quality_methods.py
@@ -77,17 +77,12 @@ from awf.db.enums import (
 )
 from awf.db.models import Workspace
 from awf.db.repositories import (
-    ResourceReservationRepository,
     WorkspaceRepository,
 )
 from awf.profiles.models import WorkspaceProfile
 from awf.runtime.validation import ValidationCoverageResult
 from awf.service.provider_recovery import (
     create_provider_recovery_attempt_row,
-)
-from awf.service.supply_chain_policy import (
-    SupplyChainPolicyRefreshResult,
-    SupplyChainPolicyRefreshService,
 )
 
 
@@ -184,40 +179,6 @@ async def _run_final_coverage_gate(
         coverage_runner=coverage_runner,
         coverage_run_kwargs=coverage_run_kwargs,
     )
-
-
-async def _parallel_worker_cpu_limit_for_workspace(
-    self: Any,
-    workspace_id: str,
-    *,
-    profile: WorkspaceProfile,
-) -> int | None:
-    if profile.validation.coverage.parallel_workers is None:
-        return None
-    async with self._session_factory() as session:
-        reservation = await ResourceReservationRepository(session).active_for_workspace(
-            workspace_id
-        )
-    if reservation is None:
-        return None
-    return max(1, int(reservation.steady_cpu))
-
-
-async def _refresh_supply_chain_policy_for_workspace(
-    self: Any,
-    *,
-    workspace_id: str,
-    command_evidence: Sequence[str],
-    changed_paths: Sequence[str],
-) -> SupplyChainPolicyRefreshResult:
-    async with self._session_factory() as session:
-        result = await SupplyChainPolicyRefreshService(session).refresh_workspace(
-            workspace_id,
-            command_evidence=command_evidence,
-            changed_paths=changed_paths,
-        )
-        await session.commit()
-        return result
 
 
 async def _verify_recovered_post_agent_commit(
